samples: framework const-correctness + sample-local include case
Second warning-cleanup wave, targeting the shared framework and the
sample-local includes the full 181-sample sweep surfaced (AlphaFog alone
did not pull these translation units in):

* include case (-Wnonportable-include-path): generic resolver now fixes
  sample-local quote includes (app.h, game.h, mesh.h, ...) against
  each file's own directory, plus any remaining SDK/framework headers --
  227 more includes across 129 files.

* XbRandName.cpp (-Wwritable-strings x126): the random-name generator's
  local literal tables and the GetVowel/GetLeadingConsonant/
  GetTrailingConsonant getters are now const WCHAR* (consumed only via
  lstrcatW, so const-clean).

* XbOnline: m_strHeader / SetHeader / local strInfo -> const WCHAR*.
* XBPerf.cpp: guard MSVC-only #pragma optimize(g,...) behind _MSC_VER
  (-Wignored-pragmas); clang cannot honor that form.
* XBStrip.cpp: compare the vertex-cache sentinel as (WORD)-1 instead of -1
  (-Wtautological-constant-out-of-range-compare).

Common/Src framework is now warning-clean. Verified representative samples
across Graphics/Networking/XboxLive3 still build.

+++ XDKSamples/Common/Src/xbstrip.cpp
@@ -1246,7 +1246,7 @@ int CVertCache::Add( int strip, int vertindex )
     int retval = 1;
 
     // If we are push one of the verts add by our strip out of the cache, return two.
-    if ( m_rgCache[m_iCachePtr] != -1 && m_rgCacheStrip[m_iCachePtr] == strip && 
+    if ( m_rgCache[m_iCachePtr] != (WORD)-1 && m_rgCacheStrip[m_iCachePtr] == strip &&
          !m_bReUsed[m_iCachePtr] )
         retval = 2;
